webapp/requirements.dev.txt
-r requirements.txt

# Main
psycopg2-binary
python-dotenv

# Development tools
django-debug-toolbar
django-extensions
ipython

# Admin
django-jazzmin

# Clean Coding
flake8
isort
black